pull resource merging into its own file
