+++ packages/honeycomb-opentelemetry-web/src/honeycomb-otel-sdk.ts
@@ -1,16 +1,13 @@
 import { WebSDK } from './base-otel-sdk';
 import { HoneycombOptions } from './types';
-import { configureHoneycombResource } from './honeycomb-resource';
-import { configureEntryPageResource } from './entry-page-resource';
-import { configureBrowserAttributesResource } from './browser-attributes-resource';
 import { configureDebug } from './honeycomb-debug';
 import { configureDeterministicSampler } from './deterministic-sampler';
 import { validateOptionsWarnings } from './validate-options';
 import { WebVitalsInstrumentation } from './web-vitals-autoinstrumentation';
 import { GlobalErrorsInstrumentation } from './global-errors-autoinstrumentation';
-import { resourceFromAttributes } from '@opentelemetry/resources';
 import { configureTraceExporters } from './composite-exporter';
 import { configureSpanProcessors } from './configure-span-processors';
+import { configureResourceAttributes } from './configure-resource-attributes';
 
 export class HoneycombWebSDK extends WebSDK {
   constructor(options?: HoneycombOptions) {
@@ -30,26 +27,10 @@ export class HoneycombWebSDK extends WebSDK {
       );
     }
 
-    //TODO: make a helper function for this?
-    let resource = resourceFromAttributes({})
-      .merge(configureEntryPageResource(options?.entryPageAttributes))
-      .merge(configureBrowserAttributesResource())
-      .merge(configureHoneycombResource());
-
-    if (options?.resource) {
-      resource = resource.merge(options.resource);
-    }
-
-    if (options?.resourceAttributes) {
-      resource = resource.merge(
-        resourceFromAttributes(options.resourceAttributes),
-      );
-    }
-
     super({
       ...options,
       instrumentations,
-      resource: resource,
+      resource: configureResourceAttributes(options),
       sampler: configureDeterministicSampler(options),
       spanProcessors: configureSpanProcessors(options),
       traceExporter: configureTraceExporters(options),
